Fix toolbar bug v2
Fix bug when toolbar was not displayed after RecyclerView reorganize and there is not enough elements anymore to enable (v2: ENOUGH) scrolling.
FYI : It is when scrolling that the toolbar displays, so if no more scrolling, no more way to display the toolbar.
V2 : When there is not enough space, scrolling is enabled but not the user can't scroll enough to display the full toolbar. Problem is FAB shows up only when the toolbar is fully visible